backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
aiosqlite==0.20.0
greenlet==3.3.2
pydantic[email]==2.9.0
pydantic-settings==2.5.2
email-validator==2.3.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==24.1.0
pillow==10.4.0
cloudinary==1.40.0
google-generativeai==0.8.3
